+++ app/devops/tests/functional/check_ins/test_functional_check_ins_create.py
@@ -6,7 +6,7 @@ from rest_framework.reverse import reverse
 
 from access.models.tenant import Tenant as Organization
 
-from devops.models.feature_flag_history import FeatureFlag
+from devops.models.feature_flag import FeatureFlag
 from devops.models.software_enable_feature_flag import SoftwareEnableFeatureFlag
 
 from itam.models.software import Software