+++ app/project_management/models/__init__.py
@@ -1,8 +1,3 @@
-from .project_milestone_history import ProjectMilestoneHistory    # pylint: disable=W0611:unused-import
-from .project_history import ProjectHistory    # pylint: disable=W0611:unused-import
-from .project_type_history import ProjectTypeHistory    # pylint: disable=W0611:unused-import
-from .project_state_history import ProjectStateHistory    # pylint: disable=W0611:unused-import
-
 from .project_notes import ProjectNotes    # pylint: disable=W0611:unused-import
 from .project_type_notes import ProjectTypeNotes    # pylint: disable=W0611:unused-import
 from .project_state_notes import ProjectStateNotes    # pylint: disable=W0611:unused-import