+++ app/settings/models/__init__.py
@@ -1,8 +1,4 @@
 from . import external_link    # pylint: disable=W0611:unused-import
 from . import user_settings    # pylint: disable=W0611:unused-import
 
-
-from .app_settings_history import AppSettingsHistory    # pylint: disable=W0611:unused-import
-from .external_link_history import ExternalLinkHistory    # pylint: disable=W0611:unused-import
-
 from .external_link_notes import ExternalLinkNotes    # pylint: disable=W0611:unused-import